Move project pamphlet properties to new file
