+++ KsApi/types/ServiceProjectWebURL.swift
@@ -13,4 +13,3 @@ extension GraphAPI.ProjectFragment: HasServiceProjectWebURL {
     self.url
   }
 }
-
